型を再配置。奥のフォルダに押し込んでいく。これはメモ辞書構築のテストを書くため
