+++ src/features/core/extract/text/index.ts
@@ -1,3 +1,4 @@
 export * from "./eventCommand";
 export * from "./mainData";
 export * from "./system";
+export * from "./types";
