+++ src/features/core/extract/types.ts
@@ -1,33 +1,11 @@
-import type {
-  MapBatchReadResult,
-  ReadSystemResult,
-  DataReadErrorItem,
-} from "@RpgTypes/fileio";
+import type { KeyValuePairEx } from "@RpgTypes/libs";
 import type { SummarizedNote, SummarizedNoteValue } from "./note";
 import type {
-  ExtractedDataBundle,
-  ExtractedCommonEventText,
-  ExtractedBattleEventText,
-  ExtractedMapTexts,
+  ExtractedRawGameDataTexts,
   ExtractedTextMainDataFinal,
+  ExtractedEventDataBundle,
+  SystemTexts,
 } from "./text";
-import type { ExtractedEventDataBundle } from "./text/eventCommand/types/bundle";
-import type { ExtractedSystemTexts, SystemTexts } from "./text/system";
-
-export interface ExtractedRawGameDataTextValue {
-  mainData: ExtractedDataBundle;
-  eventData: {
-    commonEvents: ExtractedCommonEventText[];
-    troops: ExtractedBattleEventText[];
-  };
-  mapFiles: MapBatchReadResult<ExtractedMapTexts>;
-  system: ReadSystemResult<ExtractedSystemTexts>;
-}
-
-export interface ExtractedRawGameDataTexts {
-  value: ExtractedRawGameDataTextValue;
-  errors: DataReadErrorItem[];
-}
 
 export interface RawGameDataNoteNormalization {
   nonTextNoteKeys: Set<string>;
@@ -40,4 +18,14 @@ export interface ExtractedTextFinalWithNotes<UUID>
   extends ExtractedTextMainDataFinal<UUID>, ExtractedEventDataBundle<UUID> {
   noteSummaries: SummarizedNote<SummarizedNoteValue>[];
   system: SystemTexts<UUID>;
+}
+
+export interface RuntimeDictionaryData<Hash> {
+  targetNoteKeys: string[];
+  dictionary: KeyValuePairEx<Hash, string>[];
+}
+
+export interface ReplaceAuxiliaryData<T> {
+  actorTextDictionary: KeyValuePairEx<T, string>[];
+  dictionary: RuntimeDictionaryData<T>;
 }
